pyproject.toml
[project]
name = "portfolio-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cloudinary>=1.44.2",
    "fastapi>=0.136.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.17",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langsmith>=0.8.0",
    "pygithub>=2.9.1",
    "pymongo[srv]==3.12",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.46.0",
]
